frontend/pyproject.toml
[project]
name = "chatwait-frontend"
version = "0.1.0"
description = "ChatWait frontend with Chainlit UI"
authors = [
    {name = "ChatWait Team", email = "team@chatwait.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.8.1",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.13.2",
    "black>=25.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "chainlit.*",
]
ignore_missing_imports = true